Sync2AliDisk/requirement.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.2.5
Flask==2.0.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
image==1.5.33
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==2.0.0
Jinja2==3.0.0
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==2.0.0
numpy==1.21.0
parsel==1.6.0
Pillow==8.3.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyExecJS==1.5.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pystache==0.5.4
pytz==2021.1
qrcode==7.1
queuelib==1.6.1
requests==2.25.1
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
static==1.1.1
Twisted==21.2.0
urllib3==1.26.4
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.0
zope.interface==5.4.0
